requirements.txt
camelot-py==0.11.0
Flask==3.0.3
pandas==2.2.3
pypdf==5.0.1
opencv-python=4.10.0.84